+++ aspecd/analysis.py
@@ -219,7 +219,7 @@ import aspecd.utils
 from aspecd.history import AnalysisHistoryRecord
 
 
-class AnalysisStep:
+class AnalysisStep(aspecd.utils.ToDictMixin):
     """
     Base class for analysis steps.
 
@@ -307,6 +307,7 @@ class AnalysisStep:
     """
 
     def __init__(self):
+        super().__init__()
         self.name = aspecd.utils.full_class_name(self)
         self.parameters = dict()
         self.result = None
@@ -314,6 +315,7 @@ class AnalysisStep:
         self.description = 'Abstract analysis step'
         self.comment = ''
         self.references = []
+        self._exclude_from_to_dict = ['name', 'description', 'references']
 
     def analyse(self):
         """Perform the actual analysis step on the given dataset.
@@ -462,6 +464,7 @@ class SingleAnalysisStep(AnalysisStep):
         self.description = 'Abstract single analysis step'
         self.dataset = None
         self.__kind__ = 'singleanalysis'
+        self._exclude_from_to_dict.extend(['dataset'])
 
     # pylint: disable=arguments-differ
     def analyse(self, dataset=None, from_dataset=False):
@@ -617,6 +620,7 @@ class MultiAnalysisStep(AnalysisStep):
         self.datasets = []
         self.description = 'Abstract analysis step for multiple datasets'
         self.__kind__ = 'multianalysis'
+        self._exclude_from_to_dict.extend(['datasets'])
 
     def analyse(self):
         """Perform the actual analysis on the given list of datasets.